refactor: atomise plugin
