+++ packages/portal/src/plugins/auth/constants.js
@@ -1,0 +1,2 @@
+export const PLUGIN_NAME = 'auth';
+export const NUXT_STATE_KEY = `$${PLUGIN_NAME}`;
